+++ packages/core/src/createActor.ts
@@ -121,11 +121,11 @@ export class Actor<TLogic extends AnyActorLogic>
     EmittedFrom<TLogic>
   >;
 
-  private _systemId: string | undefined;
   /** @internal */
   public _lastSourceRef?: AnyActorRef;
   /** @internal */
   public _collectedMicrosteps: AnyTransitionDefinition[] = [] as any;
+  public systemId: string | undefined;
 
   /** The globally unique process ID for this invocation. */
   public sessionId: string;
@@ -241,7 +241,7 @@ export class Actor<TLogic extends AnyActorLogic>
     // unified '@xstate.transition' event replaces '@xstate.actor'
 
     if (systemId) {
-      this._systemId = systemId;
+      this.systemId = systemId;
       this.system._set(systemId, this);
     }
 
@@ -511,8 +511,8 @@ export class Actor<TLogic extends AnyActorLogic>
     }
 
     this.system._register(this.sessionId, this);
-    if (this._systemId) {
-      this.system._set(this._systemId, this);
+    if (this.systemId) {
+      this.system._set(this.systemId, this);
     }
     this._processingStatus = ProcessingStatus.Running;